watcher/requirements.txt
slack_sdk
python-dotenv
kubernetes
pyyaml
httpx
mistralai
fastapi
uvicorn
prometheus_client

